Rename all package.json files and remove Renovate
The final versions of the slides are already built and online at the URL below.

https://work.karlhorky.com/talks/

At the moment having the package.json files in there just causes many security
warnings to be reported.

In case there should be any updates or fixes, the package.json files can always
be renamed back and work started again.